+++ frontend/src/components/Footer.jsx
@@ -1,0 +1,7 @@
+export default function Footer() {
+    return (
+        <footer>
+            Footer
+        </footer>
+    );
+}
