+++ frontend/src/components/Navbar.jsx
@@ -1,0 +1,10 @@
+import { Link } from "react-router-dom";
+
+export default function Navbar() {
+    return (
+        <nav>
+            <Link to="/">Home</Link>
+            <Link to="/about">About</Link>
+        </nav>
+    );
+}
