+++ frontend/src/pages/About.jsx
@@ -1,0 +1,3 @@
+export default function About() {
+  return <h1>About Page</h1>
+}
